Add _form partiel for create and edit template

+++ src/Controller/PinsController.php
@@ -19,13 +19,13 @@ class PinsController extends AbstractController
      */
     public function index(PinRepository $pinRepository): Response
     {
-        $pins = $pinRepository->findAll(); 
+        $pins = $pinRepository->findAll();
         return $this->render('pins/index.html.twig', [
             'pins' => $pins
         ]);
     }
 
-  
+
     /**
      * @Route("/pins/create", name="app_pins_create", methods={"GET", "POST"})
      */
@@ -35,16 +35,15 @@ class PinsController extends AbstractController
 
         $form = $this->createForm(PinType::class, $pin);
 
-            $form->handleRequest($request);
+        $form->handleRequest($request);
 
-            if ($form->isSubmitted() && $form->isValid())
-            {
-                $pin = $form->getData();
-                $em->persist($pin);
-                $em->flush();
-                
-                return $this->redirectToRoute('app_home');
-            }
+        if ($form->isSubmitted() && $form->isValid()) {
+            $pin = $form->getData();
+            $em->persist($pin);
+            $em->flush();
+
+            return $this->redirectToRoute('app_home');
+        }
 
         return $this->render('pins/create.html.twig', [
             'form' => $form->createView()
@@ -54,55 +53,51 @@ class PinsController extends AbstractController
      * @Route("/pins/{id<[0-9]+>}/edit", name="app_pins_edit", methods={"GET", "PUT"})
      *
      */
-    
-    public function edit(Request $request, Pin $pin,EntityManagerInterface $em) : Response 
+
+    public function edit(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
         $form = $this->createForm(PinType::class, $pin, [
             'method' => 'PUT'
         ]);
-        
+
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid())
-        {
+        if ($form->isSubmitted() && $form->isValid()) {
             $pin = $form->getData();
             $em->persist($pin);
             $em->flush();
-            
+
             return $this->redirectToRoute('app_home');
-            
-            
         }
         return $this->render('pins/edit.html.twig', [
             'pin' => $pin,
             'form' => $form->createView()
-            
-            ]);
-        }
-        /**
-       * @Route("/pins/{id<[0-9]+>}", name="app_pins_show", methods={"GET"})
-       */
-      public function show(Pin $pin): Response
-      {
-          return $this->render('pins/show.html.twig', compact('pin'));
-      }
-      /**
-       * Delete Pin
-       * 
-       * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
-       *
-       * @param Pin $pin
-       * @param EntityManagerInterface $em
-       * @return Response
-       */
-      public function delete(Pin $pin,EntityManagerInterface $em) : Response 
-      {
-        $em->remove($pin);
-        $em->flush();
 
-        return $this->redirectToRoute('app_home');
-
-
-      }
+        ]);
     }
-    
+    /**
+     * @Route("/pins/{id<[0-9]+>}", name="app_pins_show", methods={"GET"})
+     */
+    public function show(Pin $pin): Response
+    {
+        return $this->render('pins/show.html.twig', compact('pin'));
+    }
+    /**
+     * Delete Pin
+     * 
+     * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
+     * 
+     * @param Request $request
+     * @param Pin $pin
+     * @param EntityManagerInterface $em
+     * @return Response
+     */
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
+    {
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token'))) {
+            $em->remove($pin);
+            $em->flush();
+        }
+        return $this->redirectToRoute('app_home');
+    }
+}
